requirements.txt
Random_Word==1.0.7
pandas==0.25.1
playsound==1.2.2
Pillow==8.2.0
